requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-appconf==1.0.5
django-ckeditor==6.1.0
django-crum==0.7.9
django-imagekit==4.0.2
django-import-export==2.6.1
django-js-asset==1.2.2
djangorestframework==3.12.4
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
pilkit==2.0
Pillow==8.4.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
Unidecode==1.3.2
xlrd==2.0.1
xlwt==1.3.0
